+++ compiler/Main.hs
@@ -78,7 +78,7 @@ writeSchema _ = error "writeSchema: impossible happened."
 
 cppCodegen :: Options -> IO()
 cppCodegen options@Cpp {..} = do
-    let typeMappingAliases = maybe cppTypeMapping cppCustomAllocTypeMapping allocator
+    let typeMappingAliases = maybe cppTypeMapping (cppCustomAllocTypeMapping scoped_alloc_enabled) allocator
     let typeMapping = if type_aliases_enabled then typeMappingAliases else cppExpandAliasesTypeMapping typeMappingAliases
     concurrentlyFor_ files $ codeGen options typeMapping templates
   where
@@ -100,7 +100,7 @@ cppCodegen options@Cpp {..} = do
                         ]
     core_files = [
           reflection_h export_attribute
-        , types_h header enum_header allocator alloc_ctors_enabled type_aliases_enabled
+        , types_h header enum_header allocator alloc_ctors_enabled type_aliases_enabled scoped_alloc_enabled
         , types_cpp
         , apply_h applyProto export_attribute
         , apply_cpp applyProto