+++ compiler/Options.hs
@@ -43,6 +43,7 @@ data Options
         , grpc_enabled :: Bool
         , alloc_ctors_enabled :: Bool
         , type_aliases_enabled :: Bool
+        , scoped_alloc_enabled :: Bool
         , service_inheritance_enabled :: Bool
         }
     | Cs
@@ -99,6 +100,7 @@ cpp = Cpp
     , grpc_enabled = False &= explicit &= name "grpc" &= help "Generate gRPC definitions"
     , alloc_ctors_enabled = False &= explicit &= name "alloc-ctors" &= help "Generate constructors with allocator argument"
     , type_aliases_enabled = False &= explicit &= name "type-aliases" &= help "Generate type aliases"
+    , scoped_alloc_enabled = False &= explicit &= name "scoped-alloc" &= help "Use std::scoped_allocator_adaptor for strings and containers"
     , service_inheritance_enabled = False &= explicit &= name "enable-service-inheritance" &= help "Enable service inheritance syntax in IDL"
     } &=
     name "c++" &=